requirements.txt
python-telegram-bot
pony
Jinja2
requests